refactor: split entities into different modules
