+++ src/server/router.ts
@@ -1,30 +1,7 @@
+import { routes } from './constants/routes'
 import { RouteSegments } from './enums/route-segments.enum'
-import { RouteData } from './interfaces/route-data.interface'
+import { Route } from './interfaces/route.interface'
 import { RouteHandler } from './types/route-handler.type'
-
-export interface Route {
-  path: string
-  children: Route[] | RouteSegments.ID
-  handler?: RouteHandler
-}
-
-export const routes: Route[] = [
-  {
-    path: '/',
-    children: [
-      {
-        path: 'api',
-        children: [
-          {
-            path: 'users',
-            children: RouteSegments.ID,
-            handler: () => 'handler works!!!',
-          },
-        ],
-      },
-    ],
-  },
-]
 
 export class Router {
   private routes = routes
@@ -47,9 +24,7 @@ export class Router {
     return search(segments, this.routes)
   }
 
-  public getHandler(routeData: RouteData): RouteHandler | null {
-    const { path } = routeData
-
+  public getHandler(path: string): RouteHandler | null {
     const requiredRoute = this.searchForRoute(path)
 
     if (requiredRoute && requiredRoute.handler) {
