+++ src/server/server.ts
@@ -1,6 +1,7 @@
 import http from 'node:http'
 import url from 'node:url'
 
+import { convertToRouteData } from '../shared/utils/convert-to-route-data.util'
 import { Router } from './router'
 
 export class Server {
@@ -17,11 +18,10 @@ export class Server {
 
     this.server.on('request', (req, res) => {
       const { pathname, query } = url.parse(req.url ?? '', true)
+      const routeData = convertToRouteData(req.method ?? 'GET', pathname ?? '', query)
 
-      const handler = this.router.getHandler({ path: pathname ?? '', query, method: req.method ?? 'GET' })
-      console.log(
-        handler !== null ? handler({ path: pathname ?? '', query, method: req.method ?? 'GET' }) : 'Not found',
-      )
+      const handler = this.router.getHandler(routeData.path)
+      console.log(handler !== null ? handler(routeData) : 'Not found')
 
       res.end()
     })
