+++ src/shared/utils/convert-to-route-data.util.ts
@@ -1,0 +1,9 @@
+import { ParsedUrlQuery } from 'querystring'
+
+import { RouteData } from '../../server/interfaces/route-data.interface'
+
+export const convertToRouteData = (method: string, path: string, query: ParsedUrlQuery): RouteData => ({
+  method,
+  path,
+  query,
+})
